requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
dataclasses-json==0.6.4
deepdiff==6.7.1
Django==5.0.2
django-mysql==4.12.0
djangorestframework==3.14.0
drf-excel==2.4.0
et-xmlfile==1.1.0
idna==3.6
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
mypy-extensions==1.0.0
mysqlclient==2.2.4
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.3
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
